refactor(backend): 采用uv管理项目
动机
- 提升可维护性与一致性：以 pyproject.toml 为单一事实源，替代分散的 requirements.txt/脚本。
- 加速与可复现：uv 提供极快的解析与锁定（uv.lock），跨平台可复现构建。
- 解决 IDE “导入冒红”：src 布局让编辑器与运行时的导入路径一致，减少 Pylance 报错。
- 便于多环境与多工具：通过 [project.scripts] 暴露统一入口，CI/CD 与本地一致。

主要变更
- 依赖管理
  - 使用 uv 管理与锁定依赖：uv sync / uv add / uv remove。
  - 在 pyproject.toml 中维护运行时与 dev 分组依赖；保留镜像源配置。
- 目录与入口
  - 采用 src 布局：backend/src 为源码根；应用包为 app。
  - 新增统一入口脚本：
    - backend = "app.run:main"
    - init_db = "scripts.init_database:main"
- 文档与脚本
  - README 更新：新增 uv 使用方式与一键命令；明确 Python 版本要求。
