+++ backend/.env.example
@@ -1,0 +1,6 @@
+# 后端环境变量示例（复制为 .env 并按需修改）
+DATABASE_URL=postgresql://postgres:password@localhost:5432/achievement_platform
+SECRET_KEY=dev-secret
+JWT_SECRET_KEY=dev-jwt
+PORT=5000
+FLASK_ENV=development
