+++ backend/pyproject.toml
@@ -1,0 +1,36 @@
+[project]
+name = "achievement-backend"
+version = "0.1.0"
+description = "学生成果登记与管理系统 - 后端"
+readme = "README.md"
+requires-python = ">=3.13"
+dependencies = [
+    "flask>=3.0",
+    "flask-sqlalchemy>=3.1",
+    "SQLAlchemy>=2.0",
+    "flask-migrate>=4.0",
+    "flask-jwt-extended>=4.6",
+    "flask-cors>=4.0",
+    "flask-marshmallow>=1.2",
+    "marshmallow>=3.20",
+    "marshmallow-sqlalchemy>=0.29",
+    "psycopg2-binary>=2.9",
+    "python-dotenv>=1.0",
+    "werkzeug>=3.1.3",
+    "bcrypt>=5.0.0",
+    "pillow>=12.0.0",
+]
+
+[project.scripts]
+init_db = "scripts.init_database:main"
+backend = "app.run:main"
+
+[dependency-groups]
+dev = [
+    "pymysql>=1.1.2",
+]
+
+[[tool.uv.index]]
+url = "https://pypi.mirrors.ustc.edu.cn/simple/"
+default = true
+
