--- backend/requirements.txt
@@ -1,13 +1,0 @@
-Flask
-Flask-SQLAlchemy
-Flask-Migrate
-Flask-JWT-Extended
-Flask-CORS
-psycopg2-binary
-python-dotenv
-Werkzeug
-bcrypt
-Pillow
-marshmallow
-flask-marshmallow
-marshmallow-sqlalchemy
