+++ backend/src/app/models/__init__.py
@@ -1,6 +1,6 @@
 from .class_model import Class
 from .user import User
 from .achievement import Achievement
-from .log import AuditLog
+from .audit_log import AuditLog
 
 __all__ = ['User', 'Class', 'Achievement', 'AuditLog']
